+++ assets/js/website/listen.js
@@ -1343,11 +1343,19 @@ function updateCalendar() {
                         finalColor.red = Math.round(finalColor.red / 2);
                         finalColor.green = Math.round(finalColor.green / 2);
                         finalColor.blue = Math.round(finalColor.blue / 2);
+                        var badgeInfo;
+                        if (event.active === 2)
+                        {
+                            badgeInfo = `<span class="notification badge badge-warning shadow-2" style="font-size: 1em;">TIME CHANGED</span>`;
+                        }
+                        if (event.active === -1)
+                        {
+                            badgeInfo = `<span class="notification badge badge-danger shadow-2" style="font-size: 1em;">CANCELLED</span>`;
+                        }
                         if (event.title.startsWith("Show: "))
                         {
                             var stripped = event.title.replace("Show: ", "");
-                            var eventType = event.active > -1 ? "SHOW" : `CANCELED`;
-                            var eventClass = "primary";
+                            var eventType = "SHOW";
                             var image = `<i class="fas fa-microphone" style="font-size: 96px;"></i>`;
                             var temp = stripped.split(" - ");
                             if (temp.length === 2)
@@ -1361,8 +1369,7 @@ function updateCalendar() {
                         } else if (event.title.startsWith("Prerecord: "))
                         {
                             var stripped = event.title.replace("Prerecord: ", "");
-                            var eventType = event.active > -1 ? "PRERECORD" : `CANCELED`;
-                            var eventClass = "danger";
+                            var eventType = "PRERECORD";
                             var image = `<i class="fas fa-play-circle" style="font-size: 96px;"></i>`;
                             var temp = stripped.split(" - ");
                             if (temp.length === 2)
@@ -1376,8 +1383,7 @@ function updateCalendar() {
                         } else if (event.title.startsWith("Remote: "))
                         {
                             var stripped = event.title.replace("Remote: ", "");
-                            var eventType = event.active > -1 ? "REMOTE" : `CANCELED`;
-                            var eventClass = "purple";
+                            var eventType = "REMOTE";
                             var image = `<i class="fas fa-broadcast-tower" style="font-size: 96px;"></i>`;
                             var temp = stripped.split(" - ");
                             if (temp.length === 2)
@@ -1391,15 +1397,13 @@ function updateCalendar() {
                         } else if (event.title.startsWith("Sports: "))
                         {
                             var stripped = event.title.replace("Sports: ", "");
-                            var eventType = event.active > -1 ? "SPORTS" : `CANCELED`;
-                            var eventClass = "success";
+                            var eventType = "SPORTS";
                             var line1 = "Raider Sports";
                             var line2 = stripped;
                             var image = `<i class="fas fa-trophy" style="font-size: 96px;"></i>`;
                         } else if (event.title.startsWith("Playlist: ")) {
                             var stripped = event.title.replace("Playlist: ", "");
                             var eventType = event.active > -1 ? "PLAYLIST" : `CANCELED`;
-                            var eventClass = "info";
                             var image = `<i class="fas fa-list" style="font-size: 96px;"></i>`;
                             var temp = stripped.split(" - ");
                             if (temp.length === 2)
@@ -1413,22 +1417,20 @@ function updateCalendar() {
                         } else if (event.title.startsWith("Genre: "))
                         {
                             var stripped = event.title.replace("Genre: ", "");
-                            var eventType = event.active > -1 ? "GENRE" : `CANCELED`;
-                            var eventClass = "info";
+                            var eventType = "GENRE";
                             var line1 = "";
                             var line2 = stripped;
                             var image = `<i class="fas fa-music" style="font-size: 96px;"></i>`;
                         } else {
-                            var eventType = event.active > -1 ? "EVENT" : `CANCELED`;
-                            var eventClass = "secondary";
+                            var eventType = "EVENT";
                             var line1 = "";
                             var line2 = event.title;
                             var image = `<i class="fas fa-calendar" style="font-size: 96px;"></i>`;
                         }
                         caldata.innerHTML += `<div id="calendar-event-${event.ID}" onclick="displayEventInfo(${event.ID})" style="width: 190px; position: relative; background-color: rgb(${finalColor.red}, ${finalColor.green}, ${finalColor.blue});" class="m-2 text-white rounded shadow-8">
              <div class="p-1 text-center" style="width: 100%;">${image}
-             <span class="notification badge badge-${eventClass} shadow-2" style="font-size: 1em;">${eventType}</span>
-             <div class="m-1" style="text-align: center;"><span class="text-warning" style="font-size: 1em; text-shadow: 1px 2px 2px rgba(0,0,0,0.3);">${line1}</span><br><span style="font-size: 1.25em; text-shadow: 1px 2px 2px rgba(0,0,0,0.3);">${line2}</span><br /><span class="text-info" style="font-size: 1em; text-shadow: 1px 2px 2px rgba(0,0,0,0.3);">${moment(event.start).format("hh:mm A")} - ${moment(event.end).format("hh:mm A")}</span></div>`;
+             ${badgeInfo ? badgeInfo : ``}
+             <div class="m-1" style="text-align: center;"><span class="text-white" style="font-size: 0.8em;">${eventType}</span><br><span class="text-warning" style="font-size: 1em; text-shadow: 1px 2px 2px rgba(0,0,0,0.3);">${line1}</span><br><span style="font-size: 1.25em; text-shadow: 1px 2px 2px rgba(0,0,0,0.3);">${line2}</span><br /><span class="text-info" style="font-size: 1em; text-shadow: 1px 2px 2px rgba(0,0,0,0.3);">${moment(event.start).format("hh:mm A")} - ${moment(event.end).format("hh:mm A")}</span></div>`;
                     } catch (e) {
                         console.error(e);
                         iziToast.show({